+++ packages/components-react/projects/nextjs/pages/modal-standalone/layout.tsx
@@ -1,7 +1,8 @@
 'use client';
-import { type PropsWithChildren, useCallback, useState } from 'react';
-import { PModal, PText, PLink, PButtonGroup, PButton } from '@porsche-design-system/components-react/ssr';
+
+import { PButton, PHeading, PLink, PModal, PText } from '@porsche-design-system/components-react/ssr';
 import Link from 'next/link';
+import { type PropsWithChildren, useCallback, useState } from 'react';
 
 export default function ModalStandaloneLayout({ children }: PropsWithChildren<{}>): JSX.Element {
   const [isModalOpen, setIsModalOpen] = useState<boolean>(false);
@@ -43,14 +44,15 @@ export default function ModalStandaloneLayout({ children }: PropsWithChildren<{}
         Open Modal (without route change)
       </PButton>
       {isModalOpen && (
-        <PModal heading="Some Heading" open={isModalOpen} onDismiss={onDismiss}>
+        <PModal open={isModalOpen} onDismiss={onDismiss}>
+          <PHeading slot="header" size="large" tag="h2">
+            Some Heading
+          </PHeading>
           <PText>Some Content</PText>
-          <PButtonGroup className="footer">
-            <PButton>Save</PButton>
-            <PButton type="button" variant="secondary" icon="close" onClick={onDismiss}>
-              Close
-            </PButton>
-          </PButtonGroup>
+          <PButton slot="footer">Save</PButton>
+          <PButton slot="footer" type="button" variant="secondary" icon="close" onClick={onDismiss}>
+            Close
+          </PButton>
         </PModal>
       )}
       <main>{children}</main>