+++ packages/components-react/src/examples/LinkTileExampleHyphens.tsx
@@ -1,17 +1,17 @@
-import { PLinkTile, PSelect, PSelectOption, SelectUpdateEventDetail } from '@porsche-design-system/components-react';
+import { PLinkTile, PSelect, PSelectOption, SelectChangeEventDetail } from '@porsche-design-system/components-react';
 import { Property } from 'csstype';
 import { useState } from 'react';
 
 export const LinkTileExampleHyphensPage = (): JSX.Element => {
   const [hyphens, setHyphens] = useState<Property.Hyphens>('auto');
 
-  const onUpdate = (e: CustomEvent<SelectUpdateEventDetail>) => {
+  const onChange = (e: CustomEvent<SelectChangeEventDetail>) => {
     setHyphens(e.detail.value as Property.Hyphens);
   };
 
   return (
     <>
-      <PSelect label="Select hyphens" name="hyphens" value={hyphens} onUpdate={onUpdate}>
+      <PSelect label="Select hyphens" name="hyphens" value={hyphens} onChange={onChange}>
         <PSelectOption value="auto">style="hyphens: auto;"</PSelectOption>
         <PSelectOption value="manual">style="hyphens: manual;"</PSelectOption>
         <PSelectOption value="none">style="hyphens: none;"</PSelectOption>