+++ packages/components-react/src/examples/PinCodeExampleControlled.tsx
@@ -1,18 +1,18 @@
+import { type PinCodeChangeEventDetail, PPinCode, PText } from '@porsche-design-system/components-react';
 import { useCallback, useState } from 'react';
-import { type PinCodeUpdateEventDetail, PPinCode, PText } from '@porsche-design-system/components-react';
 
 export const PinCodeExampleControlledPage = (): JSX.Element => {
   const [value, setValue] = useState('');
   const [isComplete, setIsComplete] = useState(false);
 
-  const onUpdate = useCallback((e: CustomEvent<PinCodeUpdateEventDetail>) => {
+  const onChange = useCallback((e: CustomEvent<PinCodeChangeEventDetail>) => {
     setValue(e.detail.value);
     setIsComplete(e.detail.isComplete);
   }, []);
 
   return (
     <>
-      <PPinCode label="Some Label" value={value} onUpdate={onUpdate}></PPinCode>
+      <PPinCode label="Some Label" value={value} onChange={onChange}></PPinCode>
       <PText>Current value: {value}</PText>
       <PText>Completely filled: {isComplete.toString()}</PText>
     </>