+++ packages/components-react/src/pages/index.tsx
@@ -17,7 +17,7 @@ import { AccordionPage } from './generated/Accordion';
 import { BannerPage } from './generated/Banner';
 import { BannerBasicPage } from './generated/BannerBasic';
 import { BannerErrorPage } from './generated/BannerError';
-import { BannerNeutralPage } from './generated/BannerNeutral';
+import { BannerInfoPage } from './generated/BannerInfo';
 import { BannerNoDismissPage } from './generated/BannerNoDismiss';
 import { BannerPersistentPage } from './generated/BannerPersistent';
 import { BannerPrefixedPage } from './generated/BannerPrefixed';
@@ -25,7 +25,6 @@ import { BannerSlottedDescriptionHeadingPage } from './generated/BannerSlottedDe
 import { BannerSlottedDescriptionTitlePage } from './generated/BannerSlottedDescriptionTitle';
 import { BannerWarningPage } from './generated/BannerWarning';
 import { ButtonPage } from './generated/Button';
-import { ButtonGroupPage } from './generated/ButtonGroup';
 import { ButtonPurePage } from './generated/ButtonPure';
 import { ButtonTilePage } from './generated/ButtonTile';
 import { CanvasPage } from './generated/Canvas';
@@ -43,8 +42,6 @@ import { CanvasSidebarStartPage } from './generated/CanvasSidebarStart';
 import { CanvasTitlePage } from './generated/CanvasTitle';
 import { CarouselPage } from './generated/Carousel';
 import { CheckboxPage } from './generated/Checkbox';
-import { CheckboxWrapperPage } from './generated/CheckboxWrapper';
-import { ContentWrapperPage } from './generated/ContentWrapper';
 import { CrestPage } from './generated/Crest';
 import { DisplayPage } from './generated/Display';
 import { DividerPage } from './generated/Divider';
@@ -66,9 +63,7 @@ import { DrilldownPrefixedPage } from './generated/DrilldownPrefixed';
 import { DrilldownScrolledPage } from './generated/DrilldownScrolled';
 import { DrilldownSlottedButtonHeaderPage } from './generated/DrilldownSlottedButtonHeader';
 import { FieldsetPage } from './generated/Fieldset';
-import { FieldsetWrapperPage } from './generated/FieldsetWrapper';
 import { FlagPage } from './generated/Flag';
-import { FlexPage } from './generated/Flex';
 import { Flyout_1Page } from './generated/Flyout_1';
 import { Flyout_2Page } from './generated/Flyout_2';
 import { FlyoutBasicPage } from './generated/FlyoutBasic';
@@ -94,18 +89,14 @@ import { FlyoutHeaderFooterScrollableContentPage } from './generated/FlyoutHeade
 import { FlyoutHeaderFooterScrolledPage } from './generated/FlyoutHeaderFooterScrolled';
 import { FlyoutHeaderFooterScrolledRelativePage } from './generated/FlyoutHeaderFooterScrolledRelative';
 import { FlyoutHeaderFooterSubfooterPage } from './generated/FlyoutHeaderFooterSubfooter';
-import { FlyoutHeaderFooterSubfooterPositionLeftPage } from './generated/FlyoutHeaderFooterSubfooterPositionLeft';
 import { FlyoutHeaderFooterSubfooterPositionStartPage } from './generated/FlyoutHeaderFooterSubfooterPositionStart';
 import { FlyoutHeaderFooterSubfooterScrollableContentPage } from './generated/FlyoutHeaderFooterSubfooterScrollableContent';
 import { FlyoutHeaderFooterSubfooterScrolledPage } from './generated/FlyoutHeaderFooterSubfooterScrolled';
 import { FlyoutHeaderFooterSubfooterScrolledOverlapPage } from './generated/FlyoutHeaderFooterSubfooterScrolledOverlap';
 import { FlyoutHeaderScrollableContentPage } from './generated/FlyoutHeaderScrollableContent';
-import { FlyoutPositionLeftPage } from './generated/FlyoutPositionLeft';
 import { FlyoutPositionStartPage } from './generated/FlyoutPositionStart';
 import { FlyoutPrefixedPage } from './generated/FlyoutPrefixed';
-import { GridPage } from './generated/Grid';
 import { HeadingPage } from './generated/Heading';
-import { HeadlinePage } from './generated/Headline';
 import { IconPage } from './generated/Icon';
 import { InlineNotificationPage } from './generated/InlineNotification';
 import { InputDatePage } from './generated/InputDate';
@@ -119,11 +110,8 @@ import { InputTimePage } from './generated/InputTime';
 import { InputUrlPage } from './generated/InputUrl';
 import { LinkPage } from './generated/Link';
 import { LinkPurePage } from './generated/LinkPure';
-import { LinkSocialPage } from './generated/LinkSocial';
 import { LinkTilePage } from './generated/LinkTile';
-import { LinkTileModelSignaturePage } from './generated/LinkTileModelSignature';
 import { LinkTileProductPage } from './generated/LinkTileProduct';
-import { MarquePage } from './generated/Marque';
 import { ModalPage } from './generated/Modal';
 import { ModalBackdropBlurPage } from './generated/ModalBackdropBlur';
 import { ModalBackdropShadingPage } from './generated/ModalBackdropShading';
@@ -137,7 +125,6 @@ import { ModalGridPage } from './generated/ModalGrid';
 import { ModalPrefixedPage } from './generated/ModalPrefixed';
 import { ModalScrollablePage } from './generated/ModalScrollable';
 import { ModalSlottedHeaderPage } from './generated/ModalSlottedHeader';
-import { ModalSlottedHeadingPage } from './generated/ModalSlottedHeading';
 import { ModalStickyFooterPage } from './generated/ModalStickyFooter';
 import { ModalStretchToFullWidthPage } from './generated/ModalStretchToFullWidth';
 import { ModalWithoutDismissButtonPage } from './generated/ModalWithoutDismissButton';
@@ -173,7 +160,6 @@ import { PopoverLeftPage } from './generated/PopoverLeft';
 import { PopoverRightPage } from './generated/PopoverRight';
 import { PopoverScrolledPage } from './generated/PopoverScrolled';
 import { PopoverTopPage } from './generated/PopoverTop';
-import { RadioButtonWrapperPage } from './generated/RadioButtonWrapper';
 import { RadioGroupPage } from './generated/RadioGroup';
 import { ScrollerPage } from './generated/Scroller';
 import { SegmentedControlPage } from './generated/SegmentedControl';
@@ -198,8 +184,6 @@ import { SelectScaling_0_5Page } from './generated/SelectScaling_0_5';
 import { SelectScaling_0_8Page } from './generated/SelectScaling_0_8';
 import { SelectScaling_1Page } from './generated/SelectScaling_1';
 import { SelectScaling_2Page } from './generated/SelectScaling_2';
-import { SelectWrapperPage } from './generated/SelectWrapper';
-import { SelectWrapperOpenedScrollContextPage } from './generated/SelectWrapperOpenedScrollContext';
 import { SheetPage } from './generated/Sheet';
 import { SheetBasicPage } from './generated/SheetBasic';
 import { SheetGridPage } from './generated/SheetGrid';
@@ -218,15 +202,13 @@ import { TagPage } from './generated/Tag';
 import { TagDismissiblePage } from './generated/TagDismissible';
 import { TextPage } from './generated/Text';
 import { TextareaPage } from './generated/Textarea';
-import { TextareaWrapperPage } from './generated/TextareaWrapper';
-import { TextFieldWrapperPage } from './generated/TextFieldWrapper';
 import { TextListPage } from './generated/TextList';
 import { ToastPage } from './generated/Toast';
 import { ToastBasicPage } from './generated/ToastBasic';
 import { ToastMultilinePage } from './generated/ToastMultiline';
 import { ToastOffsetPage } from './generated/ToastOffset';
 import { ToastPrefixedPage } from './generated/ToastPrefixed';
-import { ToastStateNeutralPage } from './generated/ToastStateNeutral';
+import { ToastStateInfoPage } from './generated/ToastStateInfo';
 import { ToastStateSuccessPage } from './generated/ToastStateSuccess';
 import { TypographyCyrilPage } from './generated/TypographyCyril';
 import { TypographyFallbackPage } from './generated/TypographyFallback';
@@ -258,9 +240,9 @@ export const generatedRoutes: RouteType[] = [
     element: <BannerErrorPage />,
   },
   {
-    name: 'Banner Neutral',
-    path: '/banner-neutral',
-    element: <BannerNeutralPage />,
+    name: 'Banner Info',
+    path: '/banner-info',
+    element: <BannerInfoPage />,
   },
   {
     name: 'Banner No Dismiss',
@@ -296,11 +278,6 @@ export const generatedRoutes: RouteType[] = [
     name: 'Button',
     path: '/button',
     element: <ButtonPage />,
-  },
-  {
-    name: 'Button Group',
-    path: '/button-group',
-    element: <ButtonGroupPage />,
   },
   {
     name: 'Button Pure',
@@ -386,16 +363,6 @@ export const generatedRoutes: RouteType[] = [
     name: 'Checkbox',
     path: '/checkbox',
     element: <CheckboxPage />,
-  },
-  {
-    name: 'Checkbox Wrapper',
-    path: '/checkbox-wrapper',
-    element: <CheckboxWrapperPage />,
-  },
-  {
-    name: 'Content Wrapper',
-    path: '/content-wrapper',
-    element: <ContentWrapperPage />,
   },
   {
     name: 'Crest',
@@ -503,19 +470,9 @@ export const generatedRoutes: RouteType[] = [
     element: <FieldsetPage />,
   },
   {
-    name: 'Fieldset Wrapper',
-    path: '/fieldset-wrapper',
-    element: <FieldsetWrapperPage />,
-  },
-  {
     name: 'Flag',
     path: '/flag',
     element: <FlagPage />,
-  },
-  {
-    name: 'Flex',
-    path: '/flex',
-    element: <FlexPage />,
   },
   {
     name: 'Flyout 1',
@@ -643,11 +600,6 @@ export const generatedRoutes: RouteType[] = [
     element: <FlyoutHeaderFooterSubfooterPage />,
   },
   {
-    name: 'Flyout Header Footer Subfooter Position Left',
-    path: '/flyout-header-footer-subfooter-position-left',
-    element: <FlyoutHeaderFooterSubfooterPositionLeftPage />,
-  },
-  {
     name: 'Flyout Header Footer Subfooter Position Start',
     path: '/flyout-header-footer-subfooter-position-start',
     element: <FlyoutHeaderFooterSubfooterPositionStartPage />,
@@ -673,11 +625,6 @@ export const generatedRoutes: RouteType[] = [
     element: <FlyoutHeaderScrollableContentPage />,
   },
   {
-    name: 'Flyout Position Left',
-    path: '/flyout-position-left',
-    element: <FlyoutPositionLeftPage />,
-  },
-  {
     name: 'Flyout Position Start',
     path: '/flyout-position-start',
     element: <FlyoutPositionStartPage />,
@@ -688,19 +635,9 @@ export const generatedRoutes: RouteType[] = [
     element: <FlyoutPrefixedPage />,
   },
   {
-    name: 'Grid',
-    path: '/grid',
-    element: <GridPage />,
-  },
-  {
     name: 'Heading',
     path: '/heading',
     element: <HeadingPage />,
-  },
-  {
-    name: 'Headline',
-    path: '/headline',
-    element: <HeadlinePage />,
   },
   {
     name: 'Icon',
@@ -768,29 +705,14 @@ export const generatedRoutes: RouteType[] = [
     element: <LinkPurePage />,
   },
   {
-    name: 'Link Social',
-    path: '/link-social',
-    element: <LinkSocialPage />,
-  },
-  {
     name: 'Link Tile',
     path: '/link-tile',
     element: <LinkTilePage />,
   },
   {
-    name: 'Link Tile Model Signature',
-    path: '/link-tile-model-signature',
-    element: <LinkTileModelSignaturePage />,
-  },
-  {
     name: 'Link Tile Product',
     path: '/link-tile-product',
     element: <LinkTileProductPage />,
-  },
-  {
-    name: 'Marque',
-    path: '/marque',
-    element: <MarquePage />,
   },
   {
     name: 'Modal',
@@ -856,11 +778,6 @@ export const generatedRoutes: RouteType[] = [
     name: 'Modal Slotted Header',
     path: '/modal-slotted-header',
     element: <ModalSlottedHeaderPage />,
-  },
-  {
-    name: 'Modal Slotted Heading',
-    path: '/modal-slotted-heading',
-    element: <ModalSlottedHeadingPage />,
   },
   {
     name: 'Modal Sticky Footer',
@@ -1038,11 +955,6 @@ export const generatedRoutes: RouteType[] = [
     element: <PopoverTopPage />,
   },
   {
-    name: 'Radio Button Wrapper',
-    path: '/radio-button-wrapper',
-    element: <RadioButtonWrapperPage />,
-  },
-  {
     name: 'Radio Group',
     path: '/radio-group',
     element: <RadioGroupPage />,
@@ -1163,16 +1075,6 @@ export const generatedRoutes: RouteType[] = [
     element: <SelectScaling_2Page />,
   },
   {
-    name: 'Select Wrapper',
-    path: '/select-wrapper',
-    element: <SelectWrapperPage />,
-  },
-  {
-    name: 'Select Wrapper Opened Scroll Context',
-    path: '/select-wrapper-opened-scroll-context',
-    element: <SelectWrapperOpenedScrollContextPage />,
-  },
-  {
     name: 'Sheet',
     path: '/sheet',
     element: <SheetPage />,
@@ -1258,11 +1160,6 @@ export const generatedRoutes: RouteType[] = [
     element: <TextPage />,
   },
   {
-    name: 'Text Field Wrapper',
-    path: '/text-field-wrapper',
-    element: <TextFieldWrapperPage />,
-  },
-  {
     name: 'Text List',
     path: '/text-list',
     element: <TextListPage />,
@@ -1271,11 +1168,6 @@ export const generatedRoutes: RouteType[] = [
     name: 'Textarea',
     path: '/textarea',
     element: <TextareaPage />,
-  },
-  {
-    name: 'Textarea Wrapper',
-    path: '/textarea-wrapper',
-    element: <TextareaWrapperPage />,
   },
   {
     name: 'Toast',
@@ -1303,9 +1195,9 @@ export const generatedRoutes: RouteType[] = [
     element: <ToastPrefixedPage />,
   },
   {
-    name: 'Toast State Neutral',
-    path: '/toast-state-neutral',
-    element: <ToastStateNeutralPage />,
+    name: 'Toast State Info',
+    path: '/toast-state-info',
+    element: <ToastStateInfoPage />,
   },
   {
     name: 'Toast State Success',