+++ packages/components-react/src/styles/StylesFocusExample.tsx
@@ -1,4 +1,3 @@
-import styled from 'styled-components';
 import {
   getFocusStyle,
   gridGap,
@@ -10,6 +9,7 @@ import {
   themeLightBackgroundBase,
   themeLightPrimary,
 } from '@porsche-design-system/components-react/styles';
+import styled from 'styled-components';
 
 // Wrapper
 const Wrapper = styled.div({